chore: refatorar scripts de setup e deploy para melhor controle de permissões e segurança

+++ infra/setup/setup.sh
@@ -6,18 +6,30 @@ BASE_DIR="/opt/conexao-academica"
 REPO_URL="https://github.com/marlonzl7/conexao-academica.git"
 
 echo "Criando estrutura"
+sudo mkdir -p $BASE_DIR
+sudo chown admin:admin $BASE_DIR
+
 mkdir -p $BASE_DIR/repo
 mkdir -p $BASE_DIR/env
 
 echo "Clonando repositório"
-git clone $REPO_URL $BASE_DIR/repo
+if [ ! -d "$BASE_DIR/repo/.git" ]; then
+	git clone $REPO_URL $BASE_DIR/repo
+else
+	echo "Repositório já existe, pulando clone"
+fi
 
 echo "Copiando credenciais de exemplo"
-cp $BASE_DIR/repo/infra/env/prod.env.exemplo \
-	$BASE_DIR/env/prod.env
+if [ ! -f "$BASE_DIR/env/prod.env" ]; then
+	cp $BASE_DIR/repo/infra/env/prod.env.exemplo \
+		$BASE_DIR/env/prod.env
+else
+	echo "Arquivo prod.env já existe, pulando"
+fi
 
-echo "Adicionando permissão ao diretório do projeto"
-chmod -R 750 $BASE_DIR
+echo "Ajustando permissões"
+chmod 750 $BASE_DIR
+chmod 600 $BASE_DIR/env/prod.env
 
 echo "Insira as credenciais corretas no arquivo .prod.env"
